las_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
greenlet==3.1.1
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.14.0
pgvector==0.3.6
pydantic==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
agno==1.2.8
openai==1.54.0
anthropic==0.37.0
httpx==0.27.2
python-dotenv==1.0.1
sse-starlette==1.8.2
